brand-asset-bot/requirements.txt
# Web scraping dependencies
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78
imagehash==4.3.1

# OpenAI API for GPT integration
openai==1.3.5

# CSV and data handling
pandas==2.1.3

# Environment variables
python-dotenv==1.0.0

# User agent rotation and rate limiting
fake-useragent==1.4.0
tenacity==8.2.3

# Archive handling
rarfile==4.1

# Logging
colorlog==6.8.0
